refactor: extract composer install logic to internal/tool/composer
