+++ internal/tool/composer/composer_test.go
@@ -1,0 +1,157 @@
+// Copyright 2026 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package composer
+
+import (
+	"errors"
+	"fmt"
+	"os"
+	"path/filepath"
+	"testing"
+
+	"github.com/google/go-cmp/cmp"
+	"github.com/googleapis/librarian/internal/config"
+	"github.com/googleapis/librarian/internal/testhelper"
+)
+
+func TestInstall(t *testing.T) {
+	testhelper.RequireCommand(t, "composer")
+	for _, test := range []struct {
+		name    string
+		tools   []*config.ComposerTool
+		setup   func(t *testing.T, binDir string)
+		wantErr error
+		check   func(t *testing.T, binDir string)
+	}{
+		{
+			name: "missing repo URL",
+			tools: []*config.ComposerTool{
+				{
+					Name:    "gapic-generator-php",
+					Version: "1.0.0",
+				},
+			},
+			wantErr: ErrMissingRepo,
+		},
+		{
+			name: "success",
+			tools: []*config.ComposerTool{
+				{
+					Name:    "gapic-generator-php",
+					Version: "1.0.0",
+					Repo:    "github.com/googleapis/gapic-generator-php",
+					SHA256:  "29635b02c6e505fe31cba2f88ae999f00d2710fe1d65cb7cad521a82e7c5a518",
+				},
+			},
+			setup: func(t *testing.T, binDir string) {
+				cache := t.TempDir()
+				t.Setenv("LIBRARIAN_CACHE", cache)
+				repoDir := filepath.Join(cache, "github.com/googleapis/gapic-generator-php@1.0.0")
+				if err := os.MkdirAll(filepath.Join(repoDir, "dummy"), 0o755); err != nil {
+					t.Fatal(err)
+				}
+
+				bin := t.TempDir()
+				writeExecutable(t, filepath.Join(bin, "composer"), "#!/bin/sh\nexit 0\n")
+				t.Setenv("PATH", bin+string(os.PathListSeparator)+os.Getenv("PATH"))
+			},
+			check: func(t *testing.T, binDir string) {
+				wrapperPath := filepath.Join(binDir, "gapic-generator-php")
+				b, err := os.ReadFile(wrapperPath)
+				if err != nil {
+					t.Fatal(err)
+				}
+				repoDir := filepath.Join(os.Getenv("LIBRARIAN_CACHE"), "github.com/googleapis/gapic-generator-php@1.0.0")
+				destPath := filepath.Join(repoDir, "src", "Main.php")
+				phpPath := "php"
+				want := phpWrapperContent(phpPath, destPath)
+				if diff := cmp.Diff(want, string(b)); diff != "" {
+					t.Errorf("mismatch (-want +got):\n%s", diff)
+				}
+			},
+		},
+	} {
+		t.Run(test.name, func(t *testing.T) {
+			binDir := t.TempDir()
+			if test.setup != nil {
+				test.setup(t, binDir)
+			}
+			err := Install(t.Context(), test.tools, "php", binDir)
+			if !errors.Is(err, test.wantErr) {
+				t.Fatalf("Install() error = %v, wantErr = %v", err, test.wantErr)
+			}
+			if test.check != nil {
+				test.check(t, binDir)
+			}
+		})
+	}
+}
+
+func TestCreateBinWrapper(t *testing.T) {
+	for _, test := range []struct {
+		name        string
+		wrapperName string
+	}{
+		{
+			name:        "simple wrapper",
+			wrapperName: "foo",
+		},
+		{
+			name:        "nested wrapper",
+			wrapperName: "nested/dir/foo",
+		},
+	} {
+		t.Run(test.name, func(t *testing.T) {
+			binDir := t.TempDir()
+			destPath := "/path/to/dest"
+			content := fmt.Sprintf("#!/bin/sh\nexec %q \"$@\"\n", destPath)
+			if err := createBinWrapper(test.wrapperName, content, binDir); err != nil {
+				t.Fatal(err)
+			}
+			wrapperPath := filepath.Join(binDir, test.wrapperName)
+			b, err := os.ReadFile(wrapperPath)
+			if err != nil {
+				t.Fatal(err)
+			}
+			if diff := cmp.Diff(content, string(b)); diff != "" {
+				t.Errorf("mismatch (-want +got):\n%s", diff)
+			}
+			info, err := os.Stat(wrapperPath)
+			if err != nil {
+				t.Fatal(err)
+			}
+			perm := info.Mode().Perm()
+			if perm&0o700 != 0o700 {
+				t.Errorf("wrapper permissions = %04o, want at least 0700 (rwx) for owner", perm)
+			}
+			if perm&0o022 != 0 {
+				t.Errorf("wrapper should not be writable by group/others: %04o", perm)
+			}
+		})
+	}
+}
+
+//nolint:unparam // content is the same for all calls but keeping parameter for flexibility
+func writeExecutable(t *testing.T, path string, content string) {
+	t.Helper()
+	f, err := os.OpenFile(path, os.O_WRONLY|os.O_CREATE|os.O_EXCL, 0o755)
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer f.Close()
+	if _, err := f.WriteString(content); err != nil {
+		t.Fatal(err)
+	}
+}
